refactor(user): update UserDto to require a non-null UserProfileDto and adjust related mappers and view models to handle potential null values gracefully
